Move `TrackingPolicy` enum to `src/types`
