+++ packages/analytics-nextjs/src/lib/index.ts
@@ -4,5 +4,4 @@ export * from './getAssetClickEvent';
 export * from './getRecipientInfo';
 export * from './isPrezlyTrackingAllowed';
 export * from './stringify';
-export * from './trackingPolicy';
 export * from './urlParameters';
