+++ packages/analytics-nextjs/src/lib/isPrezlyTrackingAllowed.ts
@@ -1,7 +1,8 @@
 import type { Newsroom } from '@prezly/sdk';
 
+import { TrackingPolicy } from '../types';
+
 import { isNavigatorTrackingAllowed } from './isNavigatorTrackingAllowed';
-import { TrackingPolicy } from './trackingPolicy';
 
 /**
  * - TRUE  - user clicked "Allow"
