--- packages/analytics-nextjs/src/lib/trackingPolicy.ts
@@ -1,6 +1,0 @@
-// Pulled from `@prezly/sdk` to get rid of direct dependency requirement
-export enum TrackingPolicy {
-    DEFAULT = 'default',
-    DISABLED = 'disabled',
-    CONSENT_TO_IDENTIFY = 'consent-to-identify',
-}
