+++ packages/analytics-nextjs/src/types.ts
@@ -8,3 +8,10 @@ export interface RecipientInfo {
     id: string;
     recipient_id: string;
 }
+
+// Pulled from `@prezly/sdk` to get rid of direct dependency requirement
+export enum TrackingPolicy {
+    DEFAULT = 'default',
+    DISABLED = 'disabled',
+    CONSENT_TO_IDENTIFY = 'consent-to-identify',
+}
